Final answer not correct. Test ok
The test input data is still evaluating to the correct total.
The method has been improved to always use a list of lists rather than a list of dicts.
I don't know why this answer is different to the previous answer.
I don't know where the error is and need to do some debugging.