Problem identified with this approach.
The problem occurs when processing two pairs and a jack.
This method is not always choosing the correct card to up-value with jacks.
It is actually giving the ['J3TT3', 540] a higher rank than ['J22KK', 519].
In other words it is basing it off the 3 being higher than the 2.
It should actually be looking at the K being higher than the T.