New approach started
The computation times were too long.
It seems that each starting point will find Z rarely.
I need to find the first point where they sync.
I now want to start by finding the first 1 or 2 instances of 'Z' for each starting point.
For any given starting point, the answer will be a multiple of the length of the loop...
... plus an offset (the start point of the first full loop) for each start location.