Loop filled in to work with expanded list.
Test1_input has been modified to be more challenging.
It now does not contain the visible gap, and the method has been adapted to squeeze between the pipes.
This leaves gaps in the loop, so I'm using instructions_list to retrace the loop, filling in the gaps.
This can then be used with the same fill method as before.
However, so far this is only works for test1_input. it looks like the start of the instructions_list is missing a value.