Strings of 1s and 0s to represent combinations
My latest progress provides a representations of all possible combinations for the uncertain springs.
The number of combinations matches the prediction of the math.comb() (or "nCr") formula.
Now I need to take those strings and check them.
I think I'll do this by translating each string into the structure of the damaged_group_sizes.
If the output is the same, it is a viable permutation and the count can increase.