Now solving 2023_Day12_Part1 in a few seconds.
The old process was removing a very large number of elements from the list using the remove() method.
This is not very efficient when the list is long.
This is particularly slow because the majority of elements in the all_combinations list are not eligible.
Instead I've added the eligible elements to a new list, so fewer operations are needed.

I also combined the generate_combinations function with the function to remove ineligible elements.
Not sure if this helps too.