Literal solution attempted for 2023_Day12_Part2
It seems that it's impossible to get the right answer this way.
I would guess that since the pattern repeats, it should be possible to use the repeating pattern as the basis.

In general if you have 2 possibilities for each of 2 repeating units, you'd have 4 combinations.
If you had 3 poss for 3 units, you'd have 27 combinations.
2 poss for 3 repeating units gives 2^3 = 8 combinations
Therefore if we had 5 repeating units we would need the original number of combinations^5.
This is consistent with the fact that lines with 1 arrangement in Part1 still have 1.

I've not figured out how to account for the 4 extra ? symbols inserted in the line.
Maybe one option is to assume that all 5 repeating units had the '?', then adjust the line_possibilities downwards - not sure how to adjust.
Alternatively it might be better to assume that it's not there and then adjust the line_possibilities upwards.