Attempt for Part2. testmodule introduced
The project requirements mean that I need to evaluate the start and end condition for the problem:
I need the crucible to arrive having moved at least 4 spaces in the direction it arrives.

In this version I have also introduced a test function which I can use on other problems!