Day21_Part2 complete!
I made a silly typing mistake which was making it all more difficult than it needed to be.
I resorted back to inputting the odd or even status of the copy in question into the function, to get this over the line.
However I'm not sure if this is strictly necessary