Progress towards solving
The bricks are now "fallen", meaning the resting z heights have been calculated.
This looks correct for test data.
Next step is to figure out the number of supports for each brick.
For bricks with 1 support, that support can be disintegrated.
Bricks which are not supports can also always be disintegrated.