testinput correct
Takes an extremely long time to solve, but I don't think it's going round in circles.
Not sure how I can tell if this is infinite or not.
Either way it's impractical.
I think I might be able to speed up the way to the right answer using a heap.