Correct answer for 2023_Day23_Part2
Not an elegant solution, but by using heapq it does get the answer without exploring all possible permutations.
This is not guaranteed mathematically to work.
e.g. if there was a very short path immediately after the first input, which led to a path which was longer than all other paths combined, it would be one of the last paths to get explored despite being the longest.
However, it should work more or less well for most inputs.

The max_total_count limit is essentially an arbitrarily chosen number.
In some cases the limit would need to be increased to explore more possibilities before returning an answer, but this will take longer.