New approach
Attempting to sample the number of times a component is visited in a random loop of signals.
It gives an indication of how likely a component is to be a candidate for a wire to be cut.
This is not working yet, I may need to consider the wires rather than the components.

I can still use the component-based solution to test whether a proposed group of 3 has split the wires into separate groups or not.